# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "FinSense AI API",
  "description": "Financial analytics, portfolio optimisation, ML forecasting and RAG research API. Decision support only - not investment advice.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/funds": {
   "get": {
    "tags": [
     "funds"
    ],
    "summary": "List Funds",
    "operationId": "list_funds_api_v1_funds_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 100
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "category",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 80
        },
        {
         "type": "null"
        }
       ],
       "title": "Category"
      }
     },
     {
      "name": "asset_class",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 24
        },
        {
         "type": "null"
        }
       ],
       "title": "Asset Class"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 25,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Funds Api V1 Funds Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/funds/compare": {
   "get": {
    "tags": [
     "funds"
    ],
    "summary": "Compare",
    "operationId": "compare_api_v1_funds_compare_get",
    "parameters": [
     {
      "name": "ids",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "maxLength": 100,
       "description": "Comma-separated fund ids (2-5)",
       "title": "Ids"
      },
      "description": "Comma-separated fund ids (2-5)"
     },
     {
      "name": "period",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "1m",
        "3m",
        "6m",
        "1y",
        "3y",
        "5y",
        "max"
       ],
       "type": "string",
       "default": "3y",
       "title": "Period"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Compare Api V1 Funds Compare Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/funds/{fund_id}": {
   "get": {
    "tags": [
     "funds"
    ],
    "summary": "Fund Detail",
    "operationId": "fund_detail_api_v1_funds__fund_id__get",
    "parameters": [
     {
      "name": "fund_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Fund Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Fund Detail Api V1 Funds  Fund Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/funds/{fund_id}/nav": {
   "get": {
    "tags": [
     "funds"
    ],
    "summary": "Fund Nav",
    "operationId": "fund_nav_api_v1_funds__fund_id__nav_get",
    "parameters": [
     {
      "name": "fund_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Fund Id"
      }
     },
     {
      "name": "start",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "Start"
      }
     },
     {
      "name": "end",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "End"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Fund Nav Api V1 Funds  Fund Id  Nav Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/funds/{fund_id}/risk": {
   "get": {
    "tags": [
     "funds"
    ],
    "summary": "Fund Risk",
    "operationId": "fund_risk_api_v1_funds__fund_id__risk_get",
    "parameters": [
     {
      "name": "fund_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Fund Id"
      }
     },
     {
      "name": "period",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "1m",
        "3m",
        "6m",
        "1y",
        "3y",
        "5y",
        "max"
       ],
       "type": "string",
       "default": "3y",
       "title": "Period"
      }
     },
     {
      "name": "risk_free_rate",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number",
         "maximum": 0.2,
         "minimum": 0
        },
        {
         "type": "null"
        }
       ],
       "title": "Risk Free Rate"
      }
     },
     {
      "name": "confidence",
      "in": "query",
      "required": false,
      "schema": {
       "type": "number",
       "maximum": 0.995,
       "minimum": 0.9,
       "default": 0.95,
       "title": "Confidence"
      }
     },
     {
      "name": "start",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "Start"
      }
     },
     {
      "name": "end",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "End"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Fund Risk Api V1 Funds  Fund Id  Risk Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/funds/{fund_id}/holdings": {
   "get": {
    "tags": [
     "funds"
    ],
    "summary": "Fund Holdings",
    "operationId": "fund_holdings_api_v1_funds__fund_id__holdings_get",
    "parameters": [
     {
      "name": "fund_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Fund Id"
      }
     },
     {
      "name": "as_of",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "As Of"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Fund Holdings Api V1 Funds  Fund Id  Holdings Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/funds/{fund_id}/aum": {
   "get": {
    "tags": [
     "funds"
    ],
    "summary": "Fund Aum",
    "operationId": "fund_aum_api_v1_funds__fund_id__aum_get",
    "parameters": [
     {
      "name": "fund_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Fund Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Fund Aum Api V1 Funds  Fund Id  Aum Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/funds/{fund_id}/sip": {
   "get": {
    "tags": [
     "funds"
    ],
    "summary": "Fund Sip",
    "operationId": "fund_sip_api_v1_funds__fund_id__sip_get",
    "parameters": [
     {
      "name": "fund_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Fund Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Fund Sip Api V1 Funds  Fund Id  Sip Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/benchmarks/{benchmark_id}/series": {
   "get": {
    "tags": [
     "funds"
    ],
    "summary": "Benchmark Series",
    "operationId": "benchmark_series_api_v1_benchmarks__benchmark_id__series_get",
    "parameters": [
     {
      "name": "benchmark_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Benchmark Id"
      }
     },
     {
      "name": "start",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "Start"
      }
     },
     {
      "name": "end",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "End"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Benchmark Series Api V1 Benchmarks  Benchmark Id  Series Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/analytics/correlation": {
   "post": {
    "tags": [
     "funds"
    ],
    "summary": "Correlation",
    "operationId": "correlation_api_v1_analytics_correlation_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CorrelationRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Correlation Api V1 Analytics Correlation Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "CorrelationRequest": {
    "properties": {
     "fund_ids": {
      "items": {
       "type": "integer"
      },
      "type": "array",
      "maxItems": 12,
      "minItems": 2,
      "title": "Fund Ids"
     },
     "start": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Start"
     },
     "end": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "End"
     }
    },
    "type": "object",
    "required": [
     "fund_ids"
    ],
    "title": "CorrelationRequest"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}